requirements.txt
streamlit
groq
duckduckgo-search
PyPDF2
python-docx
Pillow
requests